Move States to own directory
